Remove dead authentication code.
